perf: don't load query builder & leaflet on home page
