+++ specifyweb/frontend/js_src/lib/components/Forms/Save.tsx
@@ -7,6 +7,7 @@ import { useId } from '../../hooks/useId';
 import { useIsModified } from '../../hooks/useIsModified';
 import { commonText } from '../../localization/common';
 import { formsText } from '../../localization/forms';
+import { smoothScroll } from '../../utils/dom';
 import { listen } from '../../utils/events';
 import { replaceKey } from '../../utils/utils';
 import { Button } from '../Atoms/Button';
@@ -27,7 +28,6 @@ import { errorHandledBy } from '../Errors/FormatError';
 import { Dialog } from '../Molecules/Dialog';
 import { hasTablePermission } from '../Permissions/helpers';
 import { userPreferences } from '../Preferences/userPreferences';
-import { smoothScroll } from '../QueryBuilder/helpers';
 import { generateMappingPathPreview } from '../WbPlanView/mappingPreview';
 import { FormContext } from './BaseResourceView';
 import { FORBID_ADDING, NO_CLONE } from './ResourceView';
