+++ specifyweb/frontend/js_src/lib/components/QueryBuilder/Wrapped.tsx
@@ -46,10 +46,11 @@ import { QueryExportButtons } from './Export';
 import { QueryFields } from './Fields';
 import { QueryFromMap } from './FromMap';
 import { QueryHeader } from './Header';
-import { smoothScroll, unParseQueryFields } from './helpers';
+import { unParseQueryFields } from './helpers';
 import { getInitialState, reducer } from './reducer';
 import type { QueryResultRow } from './Results';
 import { QueryResultsWrapper } from './ResultsWrapper';
+import { smoothScroll } from '../../utils/dom';
 import { QueryToolbar } from './Toolbar';
 
 const fetchTreeRanks = async (): Promise<true> => treeRanksPromise.then(f.true);
